Python/Cencora Development/requirements.txt
requests~=2.32.3
beautifulsoup4~=4.12.3
pydantic~=2.7.4
fastapi~=0.111.0
httpcore~=1.0.5
httpx~=0.27.0
python-dotenv~=1.0.1
tiktoken~=0.7.0
langchain~=0.2.5
azure-core~=1.30.2
azure-cosmos~=4.7.0
